requirements.txt
python-dotenv
requests
beautifulsoup4
pandas
numpy
scikit-learn
python-telegram-bot==20.7
